Do not change directory to current buffer's.
quickfixsigns will always switch/cd to the current buffer's directory
after creating the diff in the buffer's directory.

This patch will change back to the old working directory.

Additionally it uses "lcd" if a local working directory is being used.

It is adopted from VCSCommand's VCSCommandChdir /
VCSCommandChangeToCurrentFileDir [1], and it may make sense to add this
functionality to tlib instead.

1: http://repo.or.cz/w/vcscommand.git/blob/HEAD:/plugin/vcscommand.vim